Fix some build warnings
Differential Revision: https://phabricator.kde.org/D13298

+++ kmymoney/kmymoney.cpp
@@ -3855,7 +3855,7 @@ void KMyMoneyApp::slotDateChanged()
 
   // +1 is to make sure that we're already in the next day when the
   // signal is sent (this way we also avoid setting the timer to 0)
-  QTimer::singleShot((dt.secsTo(nextDay) + 1)*1000, this, SLOT(slotDateChanged()));
+  QTimer::singleShot((static_cast<int>(dt.secsTo(nextDay)) + 1)*1000, this, SLOT(slotDateChanged()));
   d->m_myMoneyView->slotRefreshViews();
 }
